Add graphql-subscriptions package and create a PubSub instance for handling message subscriptions. Publish MESSAGE_ADDED event when a new message is added and update addMessage mutation to publish the event. Add Subscription type for subscribing to the messageAdded event.
On the server side, import the PubSub class from graphql-subscriptions package and create a PubSub instance named pubSub in the resolvers.js file. Use this instance to publish the MESSAGE_ADDED event when a new message is added in the addMessage mutation. Subscription type is defined to allow clients to subscribe to the messageAdded event and receive updates when a new message is added.

This commit adds support for message subscriptions using the graphql-subscriptions package.

Please note that this analysis is based solely on the provided git diff and does not consider any context outside of it.