feat: Add useSubscription to GraphQL hooks
This commit adds the useSubscription hook from @apollo/client to the `hooks.js` file in the `client/src/lib/graphql` directory. It also includes the `messageAdded` subscription from `messageAddedSubscription` in the `queries.js` file. The useSubscription function is used to update the cache with new messages when the `messageAdded` event occurs. This ensures that the client always has the latest messages available.